requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
PyPDF2
openai
aiofiles
google-generativeai
aiohttp
